% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass_NFI.R
\name{biomass_NFI}
\alias{biomass_NFI}
\title{biomass_NFI() Function}
\usage{
biomass_NFI(data, byplot = TRUE, grpby = NULL)
}
\arguments{
\item{data}{: data}

\item{byplot}{: byplot}

\item{grpby}{: grpby}
}
\description{
This function
}
\keyword{biomass}
