% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass_temp.R
\name{biomass_temp}
\alias{biomass_temp}
\title{biomass_temp() Function}
\usage{
biomass_temp(data, byplot = TRUE, grpby = NULL, largetree = TRUE)
}
\arguments{
\item{data}{: data}

\item{byplot}{: byplot}

\item{grpby}{: grpby}

\item{largetree}{: largetree}
}
\description{
This function
}
\keyword{biomass2}
