% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass_NFI.R
\name{biomass_tsvis}
\alias{biomass_tsvis}
\title{biomass_tsvis()}
\usage{
biomass_tsvis(
  data,
  grpby = NULL,
  grpby2 = NULL,
  strat = "FORTYP_SUB",
  clusterplot = FALSE,
  largetreearea = TRUE,
  Stockedland = TRUE,
  talltree = TRUE
)
}
\arguments{
\item{data}{: A \code{list} produced by \code{\link{read_NFI}} that contains 'plot' and 'tree' data frames.}

\item{grpby}{: A character vector indicating variables from 'plot' tables for grouping. Use \code{c()} to combine multiple variables.}

\item{grpby2}{: A character vector indicating variables from 'tree' tables for grouping. Use \code{c()} to combine multiple variables.}

\item{strat}{: A character vector indicating the variable used for post-stratification. In the National Forest Inventory of Korea, it is typically used by forest type.}

\item{clusterplot}{: A logical flag indicating whether to calculate for cluster plot collectively or calculate for each subplot separately.}

\item{largetreearea}{: A logical flag indicating whether to include a large tree plot as well, or only a tree plot.}

\item{Stockedland}{: A logical flag indicating whether to include only stocked land or also include other types of land.}

\item{talltree}{: A logical flag indicating whether to include only tall trees or also shrubs.}
}
\description{
biomass_tsvis() is a function that
}
