% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_NFI.R
\name{clip_NFI}
\alias{clip_NFI}
\title{clip_NFI() Function}
\usage{
clip_NFI(NFI_point_dir, District_code)
}
\arguments{
\item{NFI_point_dir}{: dir}

\item{District_code}{: District_code}
}
\description{
This function clip_NFI
}
\examples{



}
\keyword{clip_NFI}
