% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwd_biomass_NFI.R
\name{cwd_biomass_evaluate}
\alias{cwd_biomass_evaluate}
\title{cwd_biomass_evaluate() Function}
\usage{
cwd_biomass_evaluate(data, grpby = NULL, strat = "stand", Stockedland = TRUE)
}
\arguments{
\item{data}{: data}

\item{grpby}{: grpby}

\item{strat}{: 흉고단면적/개체수}

\item{Stockedland}{: 임목지}
}
\description{
This function
}
\keyword{biomass}
