% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwd_biomass_NFI.R
\name{cwd_biomass_NFI}
\alias{cwd_biomass_NFI}
\title{cwd_biomass_NFI()}
\usage{
cwd_biomass_NFI(
  data,
  byplot = FALSE,
  grpby = NULL,
  grpby2 = NULL,
  strat = "FORTYP_SUB",
  Stockedland = TRUE
)
}
\arguments{
\item{data}{: A \code{list} generated by \code{\link{read_NFI}} that contains 'plot' and 'cwd' data frames.}

\item{byplot}{: A logical flag; whether to calculate for each plot separately or for the entire dataset.}

\item{grpby}{: A character vector; variables from 'plot' tables for grouping. Use \code{c()} to combine multiple variables.}

\item{grpby2}{: A character vector; variables from 'tree' tables for grouping. Use \code{c()} to combine multiple variables.}

\item{strat}{: A character vector; the variable used for post-stratification. In the National Forest Inventory of Korea, it is typically used by forest type.}

\item{Stockedland}{: A logical flag; whether to include only stocked land or also include other types of land.}
}
\value{
cwd tree biomass
}
\description{
cwd_biomass_NFI() is a function that
}
\details{
cwd_volume_m3_ha, var_cwd_volume_m3_ha, se_cwd_volume_m3_ha, rse_cwd_volume_m3_ha,
cwd_carbon_stock_tC_ha, var_cwd_carbon_stock_tC_ha, se_cwd_carbon_stock_tC_ha, rse_cwd_carbon_stock_tC_ha
cwd_co2_stock_tCO2_ha, var_cwd_co2_stock_tCO2_ha, se_cwd_co2_stock_tCO2_ha, rse_cwd_co2_stock_tCO2_ha
}
\examples{
\dontrun{
cwd_biomass_NFI(NFI5, grpby="map")
}

}
\references{
Son, Y., Kim, R., Lee, K., Pyo, J., Kim, S., Hwang, J., Lee, S., & Park, H. (2014). Carbon emission factors and biomass allometric equations by species in Korea. Korea Forest Research Institute.
Yim, J., Moon, G., Lee, M., Kang, J., Won, M., Ahn, E., & Jeon, J. (2021). 2020 Forest inventory of Korea. Korea Forest Research Institute.
}
