% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\docType{data}
\name{data}
\alias{data}
\title{diversity Function}
\format{
An object of class \code{data.frame} with 696062 rows and 99 columns.
}
\usage{
data
}
\arguments{
\item{data}{: data}
}
\description{
This function
}
\keyword{plot}
