% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{Descriptive_statistics}
\alias{Descriptive_statistics}
\title{Descriptive_statistics Function}
\usage{
Descriptive_statistics(data, grpby = "표본점번호")
}
\arguments{
\item{data}{: data}
}
\description{
This function
}
\keyword{plot}
