% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_NFI.R
\name{diversity_NFI}
\alias{diversity_NFI}
\title{diversity Function}
\usage{
diversity_NFI(data, grpby = "표본점번호")
}
\arguments{
\item{data}{: data}

\item{grpby}{: grpby}
}
\description{
This function
}
\keyword{plot}
