% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_NFI.R
\name{evaluate_NFI}
\alias{evaluate_NFI}
\title{evaluate_NFI() Function}
\usage{
evaluate_NFI(
  data,
  grpby = NULL,
  y = NULL,
  type = "biomass",
  output = "line",
  strat = "stand",
  clusterplot = FALSE,
  largetreearea = TRUE,
  Stockedland = TRUE,
  talltree = TRUE,
  frequency = TRUE
)
}
\arguments{
\item{data}{: data}

\item{grpby}{: grpby 시도코드; 시군구코드; 읍면동 코드}

\item{y}{: y}

\item{type}{: biomass; iv}

\item{output}{: poly; line}
}
\value{
plot data
}
\description{
This function loop function
}
