% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowNFI.R
\name{flowNFI}
\alias{flowNFI}
\title{flowNFI() Function}
\usage{
flowNFI(data, grpby = "", type = "biomass")
}
\arguments{
\item{data}{: data}

\item{grpby}{: grpby}

\item{type}{: biomass; iv}
}
\value{
merge data
}
\description{
This function loop function
}
