% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_NFI.R
\name{importancevalue_evaluate}
\alias{importancevalue_evaluate}
\title{importancevalue Function}
\usage{
importancevalue_evaluate(data)
}
\arguments{
\item{data}{: data}
}
\description{
This function
}
\keyword{plot}
