% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importancevalue_NFI.R
\name{importancevalue_NFI}
\alias{importancevalue_NFI}
\title{importancevalue Function}
\usage{
importancevalue_NFI(
  data,
  frequency = TRUE,
  clusterplot = TRUE,
  largetreearea = TRUE,
  Stockedland = TRUE,
  talltree = TRUE
)
}
\arguments{
\item{data}{: data}

\item{frequency}{: 빈도사용여부}

\item{clusterplot}{: 빈도집락부선택}

\item{largetreearea}{: 대경목조사원}

\item{Stockedland}{: 임목지}

\item{talltree}{: 교목}
}
\value{
importancevalue
}
\description{
This function
}
\keyword{plot}
