% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_NFI.R
\name{iv_NFI}
\alias{iv_NFI}
\title{iv_NFI()}
\usage{
iv_NFI(
  data,
  sp = "SP",
  frequency = TRUE,
  clusterplot = TRUE,
  largetreearea = TRUE,
  Stockedland = TRUE,
  talltree = TRUE
)
}
\arguments{
\item{data}{: A \code{list} produced by \code{\link{read_NFI}} that contains 'plot' and 'tree' data frames.}

\item{sp}{: A character value indicating the column name of tree species. e.g,. SP, Genus.}

\item{frequency}{: A logical value indicating whether to use frequency in importance calculations.}

\item{clusterplot}{: A logical value indicating whether to calculate for cluster plot collectively or calculate for each subplot separately.}

\item{largetreearea}{: A logical value indicating whether to include a large tree plot as well, or only a tree plot.}

\item{Stockedland}{: A logical value indicating whether to include only stocked land or also include other types of land.}

\item{talltree}{: A logical value indicating whether to include only tall trees or also shrubs.}
}
\value{
A \code{data.frame} that includes importance value for tree species.
}
\description{
iv_NFI() is a function that calculates the importance values of tree species based on frequency, density and coverage.
please refer to the \code{\link[BiodiversityR]{importancevalue}} function in the \pkg{BiodiversityR} package.
}
\details{
The importance value is calculated as the mean from (i) the relative frequency; (ii) the relative density; and (iii) the relative coverage. The importance value ranges between 0 and 100.
Frequency is calculated as the number of plots where a species is observed divided by the total number of survey plots. Relative frequency is calculated by dividing the frequency by the sum of the frequencies of all species, multiplied by 100 (to obtain a percentage).
Density is calculated as the total number of individuals of a species. Relative density is calculated by dividing the density by the sum of the densities of all species, multiplied by 100 (to obtain a percentage).
Coverage is calculated as the total basal area of a species. Relative dominance is calculated by dividing the dominance by the sum of the dominance of all species, multiplied by 100 (to obtain a percentage).
}
\note{
수종분류 미흡 속별로 계산해야함
}
\examples{
\dontrun{iv_NFI()}

}
\seealso{
\code{\link[BiodiversityR]{importancevalue}} for calculating the importance values.
}
