% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_NFI.R
\name{iv_tsvis}
\alias{iv_tsvis}
\title{iv_tsvis()}
\usage{
iv_tsvis(
  data,
  sp = "SP",
  frequency = TRUE,
  clusterplot = TRUE,
  largetreearea = TRUE,
  Stockedland = TRUE,
  talltree = TRUE
)
}
\arguments{
\item{data}{: A \code{list} produced by \code{\link{read_NFI}} that contains 'plot' and 'tree' data frames.}

\item{sp}{: A character value indicating the column name of tree species.}

\item{frequency}{: A logical value indicating whether to use frequency in importance calculations.}

\item{clusterplot}{: A logical value indicating whether to calculate for cluster plot collectively or calculate for each subplot separately.}

\item{largetreearea}{: A logical value indicating whether to include a large tree plot as well, or only a tree plot.}

\item{Stockedland}{: A logical value indicating whether to include only stocked land or also include other types of land.}

\item{talltree}{: A logical value indicating whether to include only tall trees or also shrubs.}
}
\value{
A \code{data.frame} that includes importance value for tree species.
}
\description{
iv_tsvis() is a function that calculates the importance values of tree species based on frequency, density and coverage.
}
\details{
dddd
}
