% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalvars.R
\docType{data}
\name{nfi_sejong}
\alias{nfi_sejong}
\title{National Forest Inventory data for Geumnam-myeon, Sejong-si, Republic of Korea for testing the function}
\format{
An object of class \code{list} of length 3.
}
\usage{
nfi_sejong
}
\description{
National Forest Inventory data for Geumnam-myeon, Sejong-si, Republic of Korea for testing the function
}
\keyword{datasets}
