% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNFI_temp.R
\name{readNFI_temp}
\alias{readNFI_temp}
\title{readNFI() Function}
\usage{
readNFI_temp(dir, district = NULL, col_all = TRUE)
}
\arguments{
\item{dir}{: dir}

\item{district}{: district}

\item{col_all}{: all column read}
}
\value{
merge data nfi
}
\description{
This function read NFI data
}
