% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_NFI.R
\name{summary_NFI}
\alias{summary_NFI}
\title{Descriptive_statistics Function}
\usage{
summary_NFI(
  data,
  grpby = NULL,
  byplot = FALSE,
  clusterplot = FALSE,
  largetreearea = TRUE,
  Stockedland = TRUE,
  talltree = TRUE
)
}
\arguments{
\item{data}{: data}

\item{grpby}{: 표본점번호}

\item{byplot}{: byplot}

\item{clusterplot}{: byplot TRUE 집락 부}

\item{largetreearea}{: 대경목조사원}

\item{Stockedland}{: 임목지}

\item{talltree}{: 교목}
}
\value{
Descriptive_statistics
}
\description{
This function
}
\keyword{plot}
