% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loopfunction.R
\name{timeloop}
\alias{timeloop}
\title{timeloop() Function}
\usage{
timeloop(data, grpby = "시군구코드", type = "biomass")
}
\arguments{
\item{data}{: data}

\item{grpby}{: grpby}

\item{type}{: biomass; importance value}
}
\value{
merge data
}
\description{
This function loop function
}
