% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsvis_NFI.R
\name{tsvis_NFI}
\alias{tsvis_NFI}
\title{tsvis_NFI() Function}
\usage{
tsvis_NFI(
  data,
  grpby = NULL,
  y = NULL,
  type = "biomass",
  output = "line",
  strat = "stand_subplot",
  clusterplot = FALSE,
  largetreearea = TRUE,
  Stockedland = TRUE,
  talltree = TRUE,
  frequency = TRUE
)
}
\arguments{
\item{data}{: data}

\item{grpby}{: grpby 시도코드; 시군구코드; 읍면동 코드}

\item{y}{: y}

\item{type}{: biomass; iv}

\item{output}{: map; line}

\item{strat}{: 흉고단면적/개체수}

\item{clusterplot}{: byplot TRUE 집락}

\item{largetreearea}{: 대경목조사원}

\item{Stockedland}{: 임목지}

\item{talltree}{: 교목}

\item{frequency}{: 빈도사용여부}
}
\value{
plot data
}
\description{
This function loop function
}
