% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.R
\name{biomass}
\alias{biomass}
\title{biomass2() Function}
\usage{
biomass(data, byplot = TRUE, grpby = NULL, largetree = TRUE)
}
\arguments{
\item{data}{: data}

\item{byplot}{: byplot}

\item{grpby}{: grpby}

\item{largetree}{: largetree}
}
\description{
This function
}
\keyword{biomass2}
