% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.R
\name{biomass}
\alias{biomass}
\title{biomass() Function}
\usage{
biomass(data, byplot = TRUE, grpby = NULL)
}
\arguments{
\item{data}{: data}

\item{byplot}{: byplot}

\item{grpby}{: grpby}
}
\description{
This function
}
\keyword{biomass}
