% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass_nfi.R
\name{biomass_nfi}
\alias{biomass_nfi}
\title{Calculates biomass of tree}
\usage{
biomass_nfi(
  data,
  byplot = FALSE,
  grpby = NULL,
  grpby2 = NULL,
  strat = "FORTYP_SUB",
  clusterplot = FALSE,
  largetreearea = TRUE,
  stockedland = TRUE,
  talltree = TRUE
)
}
\arguments{
\item{data}{: A \code{list} generated by \code{\link{read_nfi}} that contains 'plot' and 'tree' data frames.}

\item{byplot}{: A logical flag; whether to calculate for each plot separately or for the entire dataset.}

\item{grpby}{: A character vector; variables from 'plot' tables for grouping. Use \code{c()} to combine multiple variables.}

\item{grpby2}{: A character vector; variables from 'tree' tables for grouping. Use \code{c()} to combine multiple variables.}

\item{strat}{: A character vector; the variable used for post-stratification. In the National Forest Inventory of Korea, it is typically used by forest type.}

\item{clusterplot}{: A logical flag; whether to calculate for cluster plot collectively or calculate for each subplot separately.}

\item{largetreearea}{: A logical flag; whether to include a large tree plot as well, or only a tree plot.}

\item{stockedland}{: A logical flag; whether to include only stocked land or also include other types of land.}

\item{talltree}{: A logical flag; whether to include only tall trees or also shrubs.}
}
\value{
A \code{data.frame} that includes biomass for plot or study areas
}
\description{
biomass_nfi() is a function that estimates volume, aboveground biomass, biomass, carbon storage and carbon dioxide storage for each plot, the entire study area, or groups within the study area.
When calculating biomass for individual trees and plots, Users can specify whether to include large tree survey plots, to focus only on tall trees and Stocked land, to treat cluster plots as single plots.
Users can choose the criteria for post-stratification and can also group results by plot or tree characteristics.
}
\details{
Biomass is calculated using methodologies employed for national statistics.
It applies national carbon emission factors to calculate biomass at the individual tree level.
To estimate biomass per hectare at the cluster or subplot level, Users can choose whether to include only basic survey trees or both basic survey trees and large tree survey trees.
The function uses the calculated per hectare biomass at the plot level to derive annual statistics, applying the Double Sampling for Post-stratification (DSS) method.
Since only 20\% of the total plots are surveyed each year, the derived annual statistics are integrated into a single time point.
The Weighted Moving Average (WMA) method, which assigns weights to the proportion of plots collected over time, is used to calculate these statistics.
The function also estimates the variance, standard error, and relative standard error of the estimated mean values.
}
\examples{
\dontrun{
biomass_nfi(NFI5, grpby="SGG")
}

}
\references{
Son, Y., Kim, R., Lee, K., Pyo, J., Kim, S., Hwang, J., Lee, S., & Park, H. (2014). Carbon emission factors and biomass allometric equations by species in Korea. Korea Forest Research Institute.
Yim, J., Moon, G., Lee, M., Kang, J., Won, M., Ahn, E., & Jeon, J. (2021). 2020 Forest inventory of Korea. Korea Forest Research Institute.
}
