% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass_NFI.R
\name{biomass_NFI}
\alias{biomass_NFI}
\title{biomass_NFI() Function}
\usage{
biomass_NFI(
  data,
  byplot = FALSE,
  grpby = NULL,
  grpby2 = NULL,
  strat = "FORTYP_SUB",
  clusterplot = FALSE,
  largetreearea = TRUE,
  Stockedland = TRUE,
  talltree = TRUE
)
}
\arguments{
\item{data}{: data}

\item{byplot}{: byplot}

\item{grpby}{: grpby}

\item{grpby2}{: grpby}

\item{strat}{: 흉고단면적/개체수}

\item{clusterplot}{: byplot TRUE 집락}

\item{largetreearea}{: 대경목조사원}

\item{Stockedland}{: 임목지}

\item{talltree}{: 교목}
}
\value{
biomass
}
\description{
This function
}
\keyword{biomass}
