% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass_NFI.R
\name{biomass_tsvis}
\alias{biomass_tsvis}
\title{biomass_tsvis() Function}
\usage{
biomass_tsvis(
  data,
  grpby = NULL,
  strat = "stand",
  clusterplot = FALSE,
  largetreearea = TRUE,
  Stockedland = TRUE,
  talltree = TRUE
)
}
\arguments{
\item{data}{: data}

\item{grpby}{: grpby}

\item{strat}{: 흉고단면적/개체수}

\item{clusterplot}{: byplot TRUE 집락}

\item{largetreearea}{: 대경목조사원}

\item{Stockedland}{: 임목지}

\item{talltree}{: 교목}
}
\description{
This function
}
\keyword{biomass}
