% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_NFI.R
\name{clip_NFI}
\alias{clip_NFI}
\title{clip_NFI() Function}
\usage{
clip_NFI(NFI_point_dir, NFI_DF = NULL, plygn = NULL, district = NULL)
}
\arguments{
\item{NFI_point_dir}{: dir}

\item{NFI_DF}{: NFI_DF}

\item{plygn}{: plygn}

\item{District}{: District}
}
\description{
This function clip_NFI
}
\keyword{clip_NFI}
