% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwd_biomass_NFI.R
\name{cwd_biomass_NFI}
\alias{cwd_biomass_NFI}
\title{cwd_biomass_NFI() Function}
\usage{
cwd_biomass_NFI(
  data,
  byplot = FALSE,
  grpby = NULL,
  grpby2 = NULL,
  strat = "FORTYP_SUB",
  Stockedland = TRUE
)
}
\arguments{
\item{data}{: data}

\item{byplot}{: byplot}

\item{grpby}{: grpby}

\item{grpby2}{: grpby2}

\item{strat}{: 흉고단면적/개체수}

\item{Stockedland}{: 임목지}
}
\value{
cwd tree biomass
}
\description{
This function
}
\keyword{biomass}
\keyword{cwd}
\keyword{tree}
