% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwd_biomass_NFI.R
\name{cwd_biomass_NFI}
\alias{cwd_biomass_NFI}
\title{Calculates biomass of Coarse Woody Debris}
\usage{
cwd_biomass_NFI(
  data,
  byplot = FALSE,
  grpby = NULL,
  grpby2 = NULL,
  strat = "FORTYP_SUB",
  Stockedland = TRUE
)
}
\arguments{
\item{data}{: A \code{list} generated by \code{\link{read_NFI}} that contains 'plot' and 'cwd' data frames.}

\item{byplot}{: A logical flag; whether to calculate for each plot separately or for the entire dataset.}

\item{grpby}{: A character vector; variables from 'plot' tables for grouping. Use \code{c()} to combine multiple variables.}

\item{grpby2}{: A character vector; variables from 'tree' tables for grouping. Use \code{c()} to combine multiple variables.}

\item{strat}{: A character vector; the variable used for post-stratification. In the National Forest Inventory of Korea, it is typically used by forest type.}

\item{Stockedland}{: A logical flag; whether to include only stocked land or also include other types of land.}
}
\value{
A \code{data.frame} that includes biomass of CWD for plot or study areas
}
\description{
cwd_biomass_NFI() is a function that estimates volume, carbon storage and carbon dioxide storage of Coarse Woody Debris (CWD) for each plot, the entire study area, or groups within the study area.
Users can specify whether to focus only on Stocked land.
Users can choose the criteria for post-stratification and can also group results by plot or tree characteristics.
}
\details{
Biomass of CWD is calculated using methodologies employed for national statistics.
It applies national carbon emission factors to calculate biomass of CWD at the individual tree level.
Then, the function estimates biomass per hectare at the plot level and uses the calculated per hectare biomass to derive annual statistics, applying the Double Sampling for Post-stratification (DSS) method.
Since only 20\% of the total plots are surveyed each year, the derived annual statistics are integrated into a single time point.
The Weighted Moving Average (WMA) method, which assigns weights to the proportion of plots collected over time, is used to calculate these statistics.
The function also estimates the variance, standard error, and relative standard error of the estimated mean values.
}
\examples{
\dontrun{
cwd_biomass_NFI(NFI5, grpby2 = "DECAY")
}

}
\references{
Son, Y., Kim, R., Lee, K., Pyo, J., Kim, S., Hwang, J., Lee, S., & Park, H. (2014). Carbon emission factors and biomass allometric equations by species in Korea. Korea Forest Research Institute.
Yim, J., Moon, G., Lee, M., Kang, J., Won, M., Ahn, E., & Jeon, J. (2021). 2020 Forest inventory of Korea. Korea Forest Research Institute.
}
