% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwd_biomass_NFI.R
\name{cwd_biomass_NFI}
\alias{cwd_biomass_NFI}
\title{cwd_biomass_NFI()}
\usage{
cwd_biomass_NFI(
  data,
  byplot = FALSE,
  grpby = NULL,
  grpby2 = NULL,
  strat = "FORTYP_SUB",
  Stockedland = TRUE
)
}
\arguments{
\item{data}{: A \code{list} produced by \code{\link{read_NFI}} that contains 'plot' and 'cwd' data frames.}

\item{byplot}{: A logical value indicating whether to calculate for each plot separately or for the entire dataset.}

\item{grpby}{: A character value indicating variables from 'plot' tables for grouping. Use \code{c()} to combine multiple variables.}

\item{grpby2}{: A character value indicating variables from 'tree' tables for grouping. Use \code{c()} to combine multiple variables.}

\item{strat}{: A character value indicating the variable used for post-stratification. In the National Forest Inventory of Korea, it is typically used by forest type.}

\item{Stockedland}{: A logical value indicating whether to include only stocked land or also include other types of land.}
}
\value{
cwd tree biomass
}
\description{
cwd_biomass_NFI() is a function that
}
\details{
cwd_volume_m3_ha, var_cwd_volume_m3_ha, se_cwd_volume_m3_ha, rse_cwd_volume_m3_ha,
cwd_carbon_stock_tC_ha, var_cwd_carbon_stock_tC_ha, se_cwd_carbon_stock_tC_ha, rse_cwd_carbon_stock_tC_ha
}
