% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwd_biomass_NFI.R
\name{cwd_biomass_tsvis}
\alias{cwd_biomass_tsvis}
\title{cwd_biomass_tsvis()}
\usage{
cwd_biomass_tsvis(data, grpby = NULL, strat = "FORTYP_SUB", Stockedland = TRUE)
}
\arguments{
\item{data}{: A \code{list} produced by \code{\link{read_NFI}} that contains 'plot' and 'cwd' data frames.}

\item{grpby}{: A character value indicating variables from 'plot' tables for grouping. Use \code{c()} to combine multiple variables.}

\item{strat}{: A character value indicating the variable used for post-stratification. In the National Forest Inventory of Korea, it is typically used by forest type.}

\item{Stockedland}{: A logical value indicating whether to include only stocked land or also include other types of land.}
}
\description{
cwd_biomass_tsvis() is a function that
}
