% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_NFI.R
\name{diversity_NFI}
\alias{diversity_NFI}
\title{Calculates species diversity indices}
\usage{
diversity_NFI(
  data,
  sp = "SP",
  table = "tree",
  basal = FALSE,
  grpby = NULL,
  byplot = FALSE,
  clusterplot = FALSE,
  largetreearea = TRUE,
  Stockedland = TRUE,
  talltree = TRUE
)
}
\arguments{
\item{data}{: A \code{list} generated by \code{\link{read_NFI}} that contains 'plot' and one of ('tree', 'herb', 'veg', 'sapling') data frames.}

\item{sp}{: A character vector; the column name of tree species. e.g., SP, Genus.}

\item{table}{: A character vector; the target table for which diversity is to be calculated. Must be one of 'tree', 'herb', 'veg', 'sapling'.}

\item{basal}{: A logical flag; whether to calculate species diversity for trees using basal area or number of individuals.}

\item{grpby}{: A character vector; variables from 'plot' tables for grouping. Use \code{c()} to combine multiple variables.}

\item{byplot}{: A logical flag; whether to calculate for each plot separately or for the entire dataset.}

\item{clusterplot}{: A logical flag; whether to calculate for cluster plot collectively or calculate for each subplot separately.}

\item{largetreearea}{: A logical flag; whether to include a large tree plot as well, or only a tree plot.}

\item{Stockedland}{: A logical flag; whether to include only stocked land or also include other types of land.}

\item{talltree}{: A logical flag; whether to include only tall trees or also shrubs.}
}
\value{
A \code{data.frame} that includes diversity indices
}
\description{
diversity_NFI() is a function that calculates species richness, evenness and the Shannon and Simpson diversity indices for each plot, the entire study area, or groups within the study area.
It can calculate diversity at the species or genus level for tables such as trees and saplings.
please refer to the \code{\link[vegan]{diversity}} function in the \pkg{vegan} package.
Users can specify whether to include large tree survey plots, to focus only on tall trees and Stocked land, and to treat cluster plots as single plots.
Users can also group diversity indices by plot characteristics.
}
\details{
Species richness refers to the total number of species surveyed.
The Shannon-Wiener index, a species diversity index, is the sum of the proportions of the number of individuals or the basal area of a specific species relative to the total number of individuals or the total basal area surveyed.
The Gini-Simpson index is calculated by subtracting Simpson's index from 1, and it ranges between 0 and 1. The value increases as diversity increases.
Species evenness, is calculated by dividing the species diversity by the maximum species diversity, and it ranges between 0 and 1. It has a maximum value of 1 when all species are evenly distributed.
The function also estimates the mean and standard error of diversity.
}
\note{
The 'herb', 'veg', and 'sapling' tables may contain a lot of errors, so please consider whether to use them.
}
\examples{
\dontrun{
diversity_NFI(NFI5, sp="SP", table="tree", basal=TRUE)
}

}
\references{
Shannon, C. E. (1948). A mathematical theory of communication. The Bell System Technical Journal, 27(3), 379–423.
Simpson, E. H. (1949). Measurement of diversity. Nature, 163(4148), 688–688.
Pielou, E. C. (1966). The measurement of diversity in different types of biological collections. Journal of Theoretical Biology, 13, 131–144.
}
\seealso{
\code{\link[vegan]{diversity}} for calculating the Shannon and Simpson diversity indices.
}
