% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_NFI.R
\name{diversity_NFI}
\alias{diversity_NFI}
\title{diversity_NFI()}
\usage{
diversity_NFI(
  data,
  sp = "SP",
  grpby = NULL,
  type = "tree",
  byplot = FALSE,
  basal = FALSE,
  clusterplot = FALSE,
  largetreearea = TRUE,
  Stockedland = TRUE,
  talltree = TRUE
)
}
\arguments{
\item{data}{: A \code{list} produced by \code{\link{read_NFI}} that contains 'plot' and one of ('tree', 'herb', 'veg', 'sapling') data frames.}

\item{sp}{: A character value indicating the column name of tree species. e.g., SP, Genus.}

\item{grpby}{: A character value indicating variables from 'plot' tables for grouping. Use \code{c()} to combine multiple variables.}

\item{type}{: Use \code{c()} to combine multiple variables.}

\item{byplot}{: A logical value indicating whether to calculate for each plot separately or for the entire dataset.}

\item{basal}{: A logical value indicating whether to calculate species diversity for trees using basal area or number of individuals.}

\item{clusterplot}{: A logical value indicating whether to calculate for cluster plot collectively or calculate for each subplot separately.}

\item{largetreearea}{: A logical value indicating whether to include a large tree plot as well, or only a tree plot.}

\item{Stockedland}{: A logical value indicating whether to include only stocked land or also include other types of land.}

\item{talltree}{: A logical value indicating whether to include only tall trees or also shrubs.}
}
\value{
A \code{data.frame} that includes diversity indices or numbers of species. standard error
}
\description{
diversity_NFI() is a function that calculates species richness, evenness and the Shannon and Simpson diversity indices for plot or study areas.
please refer to the \code{\link[vegan]{diversity}} function in the \pkg{vegan} package.
}
\details{
Species richness
Shannon-Wiener's index
Simpson's index
Evenness
entire dataset은 mean, standard error
}
\note{
식생데이터 데이터 이상함
}
\seealso{
\code{\link[vegan]{diversity}} for calculating the Shannon and Simpson diversity indices.
}
