% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_NFI.R
\name{diversity_NFI}
\alias{diversity_NFI}
\title{diversity Function}
\usage{
diversity_NFI(
  data,
  sp = "SP",
  grpby = NULL,
  type = "tree",
  byplot = FALSE,
  basal = FALSE,
  clusterplot = FALSE,
  largetreearea = TRUE,
  Stockedland = TRUE,
  talltree = TRUE
)
}
\arguments{
\item{data}{: data}

\item{sp}{: sp}

\item{grpby}{: grpby}

\item{type}{: grpby}

\item{byplot}{: byplot}

\item{basal}{: 흉고단면적/개체수}

\item{clusterplot}{: byplot TRUE 집락 부}

\item{largetreearea}{: 대경목조사원}

\item{Stockedland}{: 임목지}

\item{talltree}{: 교목}
}
\value{
diversity
}
\description{
This function
}
\keyword{plot}
