% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_nfi.R
\name{filter_nfi}
\alias{filter_nfi}
\title{filter_nfi()}
\usage{
filter_nfi(data, expr_texts, hier = T)
}
\arguments{
\item{data}{: A \code{list} generated by \code{\link{read_nfi}}}

\item{expr_texts}{: expr_texts}

\item{hier}{: hierarchy}
}
\value{
A \code{list}
}
\description{
filter_nfi() is a function that
}
\details{
dddd
}
\examples{
\dontrun{
filter_nfi(NFI5, )
}

}
