% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_NFI.R
\name{filter_NFI}
\alias{filter_NFI}
\title{filter_NFI()}
\usage{
filter_NFI(data, expr_texts, hier = T)
}
\arguments{
\item{data}{: A \code{list} generated by \code{\link{read_NFI}}}

\item{expr_texts}{: A logical flag indicating whether to load a standing tree survey table.}

\item{hier}{: A logical flag indicating whether to load a standing tree survey table.}
}
\value{
dataframe
}
\description{
filter_NFI() is a function that
}
\details{
dddd
}
\examples{
\dontrun{
filter_NFI(NFI5)
}

}
