% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_NFI.R
\name{filter_NFI}
\alias{filter_NFI}
\title{filter_NFI()}
\usage{
filter_NFI(data, expr_texts, hier = T)
}
\arguments{
\item{data}{: A \code{list} generated by \code{\link{read_NFI}}}

\item{expr_texts}{: expr_texts}

\item{hier}{: hierarchy}
}
\value{
A \code{list}
}
\description{
filter_NFI() is a function that
}
\details{
dddd
}
\examples{
\dontrun{
filter_NFI(NFI5, )
}

}
