% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowNFI.R
\name{flowNFI}
\alias{flowNFI}
\title{flowNFI() Function}
\usage{
flowNFI(data, grpby = "", y = NULL, type = "biomass", output = "line")
}
\arguments{
\item{data}{: data}

\item{grpby}{: grpby 시도코드; 시군구코드; 읍면동 코드}

\item{type}{: biomass; iv}

\item{output:}{poly; line}
}
\value{
merge data
}
\description{
This function loop function
}
