% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_NFI.R
\name{iv_NFI}
\alias{iv_NFI}
\title{Calculates the importance values}
\usage{
iv_NFI(
  data,
  sp = "SP",
  frequency = TRUE,
  clusterplot = TRUE,
  largetreearea = TRUE,
  Stockedland = TRUE,
  talltree = TRUE
)
}
\arguments{
\item{data}{: A \code{list} generated by \code{\link{read_NFI}} that contains 'plot' and 'tree' data frames.}

\item{sp}{: A character vector; the column name of tree species. e.g,. SP, Genus.}

\item{frequency}{: A logical flag; whether to use frequency in importance calculations.}

\item{clusterplot}{: A logical flag; whether to calculate for cluster plot collectively or calculate for each subplot separately.}

\item{largetreearea}{: A logical flag; whether to include a large tree plot as well, or only a tree plot.}

\item{Stockedland}{: A logical flag; whether to include only stocked land or also include other types of land.}

\item{talltree}{: A logical flag; whether to include only tall trees or also shrubs.}
}
\value{
A \code{data.frame} that includes importance value for tree species.
}
\description{
iv_NFI() is a function that calculates the importance values of tree species based on frequency, density and coverage.
You can specify whether to include frequency in the calculation.
please refer to the \code{\link[BiodiversityR]{importancevalue}} function in the \pkg{BiodiversityR} package.
You can also specify whether to include large tree survey plots, to focus only on tall trees and Stocked land, and to treat cluster plots as single plots.
}
\details{
The importance value (ranging from 0 to 100) is calculated as the mean of (i) relative frequency, (ii) relative density, and (iii) relative coverage.
Frequency is the number of plots where a species is observed divided by the total survey plots. Relative frequency is frequency divided by the sum of all species' frequencies, multiplied by 100.
Density is the total number of individuals of a species. Relative density is density divided by the sum of all species' densities, multiplied by 100.
Coverage is the total basal area of a species. Relative coverage is coverage divided by the sum of all species' coverage, multiplied by 100.
}
\note{
Species classification may be incomplete, so it might be worth considering calculating importance by genus rather than species.
Since the frequencies of each species may be identical across the nation, it may be desirable to exclude frequency from the importance calculation.
}
\examples{
\dontrun{
iv_NFI(NFI5, sp="Genus")
}

}
\references{
Curtis, J. T. & McIntosh, R. P. (1951). An upland forest continuum in the prairie-forest border region of Wisconsin. Ecology, 32(3), 476–496.
}
\seealso{
\code{\link[BiodiversityR]{importancevalue}} for calculating the importance values.
}
