% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_NFI.R
\name{iv_tsvis}
\alias{iv_tsvis}
\title{importancevalue Function}
\usage{
iv_tsvis(
  data,
  sp = "SP",
  frequency = TRUE,
  clusterplot = TRUE,
  largetreearea = TRUE,
  Stockedland = TRUE,
  talltree = TRUE
)
}
\arguments{
\item{data}{: data}

\item{sp}{: sp}

\item{frequency}{: 빈도사용여부}

\item{clusterplot}{: 빈도집락부선택}

\item{largetreearea}{: 대경목조사원}

\item{Stockedland}{: 임목지}

\item{talltree}{: 교목}
}
\description{
This function
}
\keyword{plot}
