% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_NFI.R
\name{read_NFI}
\alias{read_NFI}
\title{read_NFI() Function}
\usage{
read_NFI(dir, district = NULL)
}
\arguments{
\item{dir}{: dir}

\item{district}{: district}
}
\value{
merge data nfi
}
\description{
This function read NFI data
}
\examples{
read_NFI("D:/y2021/y202101/rNFI/NFI/NFI7/", district = "\uc804\ub77c\ub0a8\ub3c4") 
}
