% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_NFI.R
\name{read_NFI}
\alias{read_NFI}
\title{Reads Korean National Forest Inventory}
\usage{
read_NFI(dir, district = NULL, tables = c("tree", "cwd"))
}
\arguments{
\item{dir}{: A character vector indicating directory of NFI files.}

\item{district}{: A character vector indicating the district's Korean name within levels such as sido, sigungu, or eupmyondong. If \code{NULL}, the entire dataset is loaded.}

\item{tables}{: A character vector indicating names of specific tables to be imported. Use \code{c()} to combine multiple variables. e.g., \verb{c('tree', 'cwd', 'stump', herb', 'veg', 'sapling', 'soil')}.}
}
\value{
A \code{data.frame} containing the loaded and transformed NFI data, structured for easy analysis. Columns and structure depend on the survey tables loaded.
}
\description{
read_NFI() is a function that reads Korean National Forest Inventory (NFI).
Loads the annual NFI file from the local computer.
And change the data to an easy-to-analyze format and perform integrity verification.
NFI is downloaded from \url{https://kfss.forest.go.kr/stat/}.
}
\details{
This functionality performs integrity validation on data provided by the Korea Forest Service, based on the database of subplot and species.
It corrects errors in the administrative regions of subplots and in the classification of species as coniferous or deciduous.
The existing data, which only provides species names, has been augmented with the Korean and English names of families and genera, as well as the scientific names of species.
The classification of species into coniferous and broadleaf categories, as well as plant taxonomy, follows the standards of the Korean Plant Names Index Committee of the Korea National Arboretum \url{http://www.nature.go.kr/kpni/index.do}.
Additionally, this functionality calculates the forest type, dominant species, and dominant species percentage for each subplot and cluster plot.
}
\note{
To download subsets of the annual NFI file manually, go online to the Korea Forest Service Forestry Statistics Platform (\url{https://kfss.forest.go.kr/stat/}), downloaded .zip files, and extract them.
Load the data \code{rNFI::col_name} to find out the Korean and English names of the column names.
}
\examples{
\dontrun{
read_NFI("D:/NFI/NFI5", district=NULL, tables=c("tree", "cwd"))
}

}
