% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNFI.R
\name{readNFI}
\alias{readNFI}
\title{readNFI() Function}
\usage{
readNFI(dir, district = NULL, col_all = FALSE)
}
\arguments{
\item{dir}{: dir}

\item{district}{: district}

\item{col_all}{: all column read}
}
\value{
merge data nfi
}
\description{
This function read NFI data
}
\examples{
readNFI("C:/SYP/y2021/y202101/rNFI/data/NFI/NFI6/", district = "\uc804\ub77c\ub0a8\ub3c4 \uc2e0\uc548\uad70") 
}
