% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{summary_NFI}
\alias{summary_NFI}
\title{Descriptive_statistics Function}
\usage{
summary_NFI(data, grpby = NULL)
}
\arguments{
\item{data}{: data}

\item{grpby}{: 표본점번호}
}
\description{
This function
}
\keyword{plot}
