% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_nfi.R
\name{summary_nfi}
\alias{summary_nfi}
\title{Calculates descriptive statistics}
\usage{
summary_nfi(
  data,
  grpby = NULL,
  byplot = FALSE,
  clusterplot = FALSE,
  largetreearea = TRUE,
  stockedland = TRUE,
  talltree = TRUE
)
}
\arguments{
\item{data}{: A \code{list} generated by \code{\link{read_nfi}} that contains 'plot' and 'tree' data frames.}

\item{grpby}{: A character vector; variables from 'plot' tables for grouping. Use \code{c()} to combine multiple variables.}

\item{byplot}{: A logical flag; whether to calculate for each plot separately or for the entire dataset.}

\item{clusterplot}{: A logical flag; whether to calculate for cluster plot collectively or calculate for each subplot separately.}

\item{largetreearea}{: A logical flag; whether to include a large tree plot as well, or only a tree plot.}

\item{stockedland}{: A logical flag; whether to include only stocked land or also include other types of land.}

\item{talltree}{: A logical flag; whether to include only tall trees or also shrubs.}
}
\value{
A \code{data.frame} that includes summary statistics for plot or study areas
}
\description{
summary_nfi() is a function that calculates descriptive statistics for each plot, the entire study area, or groups within the study area.
Users can specify whether to include large tree survey plots, to focus only on tall trees and Stocked land, and to treat cluster plots as single plots.
}
\details{
The summary_nfi function provides statistics related to plots, including the number of cluster plots, subplots, subplots with large trees (≥30cm) observed in large tree plots, and subplots where large trees were observed.
It also offers statistics for trees, such as the number of individuals, large trees, dominant trees, and species.
Additionally, this function estimates the mean and standard error for tree measurements and stand statistics: DBH (Diameter at Breast Height), height, height of dominant trees, number of individuals per hectare, basal area per hectare, and volume per hectare.
}
\examples{
\dontrun{
summary_nfi(NFI5, grpby = "AGE_CLS")
}

}
