% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_nfi.R
\name{summary_nfi}
\alias{summary_nfi}
\title{Calculate descriptive statistics for National Forest Inventory Data}
\usage{
summary_nfi(
  data,
  grpby = NULL,
  byplot = FALSE,
  clusterplot = FALSE,
  largetreearea = TRUE,
  stockedland = TRUE,
  talltree = TRUE
)
}
\arguments{
\item{data}{: A \code{list} containing 'plot' and 'tree' data frames, typically generated by \code{\link{read_nfi}}.}

\item{grpby}{: A character vector; specifies variables from 'plot' table to use for grouping. Use \code{c()} to combine multiple variables.}

\item{byplot}{: A logical flag (default FALSE); if TRUE, calculates statistics for each plot separately. If FALSE, calculates for the entire dataset.}

\item{clusterplot}{: A logical flag (default FALSE); if TRUE, treats each cluster plot as a single unit. If FALSE, calculates for each subplot separately.}

\item{largetreearea}{: A logical flag (default TRUE); if TRUE, includes large tree survey plots in the analysis. If FALSE, only uses standard tree plots.}

\item{stockedland}{: A logical flag (default TRUE); if TRUE, includes only stocked land. If FALSE, includes all land types.}

\item{talltree}{: A logical flag (default TRUE); if TRUE, includes only tall trees. If FALSE, includes both trees and shrubs.}
}
\value{
A \code{data.frame} that includes summary statistics.
The structure depends on the input parameters:
\itemize{
\item If \code{byplot = TRUE}, each row represents a plot.
\item If \code{byplot = FALSE}, each row represents the entire dataset or a group specified by \code{grpby}
}
}
\description{
summary_nfi() is a function that calculates comprehensive descriptive statistics for National Forest Inventory (NFI) data.
It can provide summaries for individual plots, the entire study area, or specific groups within the study area.
Users have flexibility in specifying data inclusion criteria and analysis levels.
Users can specify whether to include large tree survey plots, to focus only on tall trees and Stocked land, and to treat cluster plots as single plots.
}
\details{
The summary_nfi function provides statistics related to plots, including the number of cluster plots, subplots, subplots with large trees (≥30cm) observed in large tree survey plots, and subplots where large trees were observed.
It also offers statistics for trees, such as the number of individuals, large trees, dominant trees, and species.
Additionally, this function estimates the mean and standard error for tree measurements and stand statistics: DBH (Diameter at Breast Height), height, height of dominant trees, number of individuals per hectare, basal area per hectare, and volume per hectare.
}
\examples{
\dontrun{
# Basic usage
summary_stats <- summary_nfi(NFI5)

# Summarize by the group, including all land types
grouped_stats <- summary_nfi(NFI5, grpby = "AGE_CLS", stockedland = FALSE)

# Summarize by individual plots, including both trees and shrubs 
plot_summaries <- summary_nfi(NFI5, byplot = TRUE, talltree=FALSE)
}

}
