% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsvis_NFI.R
\name{tsvis_NFI}
\alias{tsvis_NFI}
\title{tsvis_NFI()}
\usage{
tsvis_NFI(
  data,
  grpby = NULL,
  y = NULL,
  type = "biomass",
  output = "line",
  strat = "FORTYP_SUB",
  clusterplot = FALSE,
  largetreearea = TRUE,
  Stockedland = TRUE,
  talltree = TRUE,
  frequency = TRUE
)
}
\arguments{
\item{data}{: A \code{list} produced by \code{\link{read_NFI}} that contains 'plot' and 'tree' data frames.}

\item{grpby}{: grpby 시도코드; 시군구코드; 읍면동 코드}

\item{y}{: y}

\item{type}{: biomass; iv; cwd}

\item{output}{: map; line; table; bar}

\item{strat}{: A character value indicating the variable used for post-stratification. In the National Forest Inventory of Korea, it is typically used by forest type.}

\item{clusterplot}{: A logical value indicating whether to calculate for cluster plot collectively or calculate for each subplot separately.}

\item{largetreearea}{: A logical value indicating whether to include a large tree plot as well, or only a tree plot.}

\item{Stockedland}{: A logical value indicating whether to include only stocked land or also include other types of land.}

\item{talltree}{: A logical value indicating whether to include only tall trees or also shrubs.}

\item{frequency}{: A logical value indicating whether to use frequency in importance calculations.}
}
\value{
plot data
}
\description{
tsvis_NFI() is a function that
}
\details{
This function loop function
}
